Add a module for initialisations
