+++ modules/Initialisations.js
@@ -1,0 +1,14 @@
+// Define page navigation
+const hamburger = document.querySelector('.hamburger');
+const closeIcon = document.querySelector('#close-icon');
+const dropDownNav = document.querySelector('.navigation-menu');
+const listNav = document.querySelector('#list-nav');
+const addNewNav = document.querySelector('#add-new-nav');
+const contactNav = document.querySelector('#contact-nav');
+
+// Define page sections
+const viewBooksSection = document.querySelector('.view-books');
+const addBookFormSection = document.querySelector('.add-book-form');
+const newContactSection = document.querySelector('.new-contact-info');
+
+export { hamburger, closeIcon, dropDownNav, listNav, addNewNav, contactNav, viewBooksSection, addBookFormSection, newContactSection };
