+++ script.js
@@ -1,19 +1,7 @@
-/* eslint-disable max-classes-per-file */
-// Define page navigation
-const hamburger = document.querySelector('.hamburger');
-const closeIcon = document.querySelector('#close-icon');
-const dropDownNav = document.querySelector('.navigation-menu');
-const listNav = document.querySelector('#list-nav');
-const addNewNav = document.querySelector('#add-new-nav');
-const contactNav = document.querySelector('#contact-nav');
-/* eslint-disable max-classes-per-file */
 
-// Define page sections
-const viewBooksSection = document.querySelector('.view-books');
-const addBookFormSection = document.querySelector('.add-book-form');
-const newContactSection = document.querySelector('.new-contact-info');
+// Import page navigation and page sections initialisations
+import { hamburger, closeIcon, dropDownNav, listNav, addNewNav, contactNav, viewBooksSection, addBookFormSection, newContactSection } from './modules/Initialisations.js';
 
-/* Add event listeners */
 // List event listener
 listNav.addEventListener('click', () => {
   viewBooksSection.classList.remove('hide');
